+++ .github/workflows/build.yml
@@ -159,7 +159,7 @@ jobs:
           if [ "${{ matrix.linking }}" = "dynamic" ]; then
             FEATURES="${FEATURES},${FEATURE_SET},dynamic"
           elif [ "${{ matrix.linking }}" = "static" ]; then
-            FEATURES="${FEATURES},${FEATURE_SET},ossl400"
+            FEATURES="${FEATURES},${FEATURE_SET},ossl-sys,ossl400"
           elif [ "${{ matrix.linking }}" = "fips" ]; then
             FEATURES="${FEATURES},fips"
           elif [ "${{ matrix.linking }}" = "openssl-sys" ]; then
