+++ ossl-sys/README.md
@@ -1,0 +1,7 @@
+# ossl-sys
+
+A crate providing low-level Rust bindings to the OpenSSL library.
+
+`ossl-sys` was spun off from the `ossl` crate to provide unsafe low-level
+FFI bindings. It is intended as a foundational building block for
+higher-level cryptographic libraries rather than for direct application use.
