+++ ossl-sys/src/lib.rs
@@ -1,0 +1,9 @@
+// Copyright 2025 Simo Sorce
+// See LICENSE.txt file for terms
+
+#![allow(non_upper_case_globals)]
+#![allow(non_camel_case_types)]
+#![allow(dead_code)]
+#![allow(non_snake_case)]
+#![allow(unnecessary_transmutes)]
+include!(concat!(env!("OUT_DIR"), "/ossl_bindings.rs"));
