+++ ossl/README.md
@@ -10,7 +10,6 @@ the high-level EVP interface and the `OSSL_PARAM` mechanism for parameter
 handling. It avoids legacy APIs to promote a more consistent and
 forward-looking approach to using OpenSSL from Rust.
 
-This crate provides unsafe FFI bindings and some thin, safe wrappers around
-common OpenSSL objects and patterns. It is intended as a foundational
-building block for higher-level cryptographic libraries rather than for
-direct application use.
+This crate provides some thin, safe wrappers around common OpenSSL objects
+and patterns. It is intended as a foundational building block for higher-level
+cryptographic libraries rather than for direct application use.
