+++ ossl/src/lib.rs
@@ -8,14 +8,14 @@
 /// Part of this module is automatically generated by bindgen from the OpenSSL
 /// Headers and includes a selection of functions and other items needed to
 /// access the libcrypto/libfips functions needed.
-#[cfg(not(feature = "openssl-sys"))]
+#[cfg(feature = "ossl-sys")]
 pub mod bindings {
     #![allow(non_upper_case_globals)]
     #![allow(non_camel_case_types)]
     #![allow(dead_code)]
     #![allow(non_snake_case)]
     #![allow(unnecessary_transmutes)]
-    include!(concat!(env!("OUT_DIR"), "/ossl_bindings.rs"));
+    pub use ossl_sys::*;
 }
 #[cfg(feature = "openssl-sys")]
 pub mod bindings {
@@ -53,27 +53,14 @@ pub mod fips;
 /// This returns a triplet of u8 values representing Major, Minor, and
 /// Patch version.
 pub fn api_level() -> (u8, u8, u8) {
-    #[cfg(not(feature = "openssl-sys"))]
-    {
-        // This is the 0xMMNNPP (Major, miNor, Patch) scheme
-        let version = OPENSSL_API_LEVEL;
-        let patch: u8 = (version & 0xff) as u8;
-        let minor: u8 = ((version & 0xff00) >> 8) as u8;
-        let major: u8 = ((version & 0xff0000) >> 16) as u8;
-        (major, minor, patch)
-    }
+    // This is the 0xMNN00PP0 (Major, miNor, Patch) scheme
+    let v = env!("OPENSSL_VERSION_NUMBER");
+    let version = u64::from_str_radix(&v, 16).unwrap();
 
-    #[cfg(feature = "openssl-sys")]
-    {
-        // This is the 0xMNN00PP0 (Major, miNor, Patch) scheme
-        let v = env!("DEP_OPENSSL_VERSION_NUMBER");
-        let version = u64::from_str_radix(&v, 16).unwrap();
-
-        let patch: u8 = ((version & 0xff0) >> 4) as u8;
-        let minor: u8 = ((version & 0xff00000) >> 20) as u8;
-        let major: u8 = ((version & 0xff0000000) >> 28) as u8;
-        (major, minor, patch)
-    }
+    let patch: u8 = ((version & 0xff0) >> 4) as u8;
+    let minor: u8 = ((version & 0xff00000) >> 20) as u8;
+    let major: u8 = ((version & 0xff0000000) >> 28) as u8;
+    (major, minor, patch)
 }
 
 /// Securely zeroizes a memory slice using `OPENSSL_cleanse`.
